+++ app/route.js
@@ -25,4 +25,12 @@ async function router(fastify) {
     fastify.register(require('./routes/isonline/report'), { prefix: `${rootPrefix}/isonline/report`, logger: true });
     fastify.register(require('./routes/isonline/moph'), { prefix: `${rootPrefix}/moph`, logger: true });
     fastify.register(require('./routes/isonline/ops'), { prefix: `${rootPrefix}/ops`, logger: true });
+    fastify.register(require('./routes/pcc/index'), { prefix: `${rootPrefix}/pcc`, logger: true });
+    fastify.setNotFoundHandler((request, reply) => {
+        reply.send({
+            statusCode: 404,
+            error: 'Not Found',
+            message: `Route ${request.method}:${request.url} not found`
+        });
+    });
 }