+++ app/routes/pcc/index.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const HttpStatus = require("http-status-codes");
+const http_status_codes_1 = require("http-status-codes");
 var crypto = require('crypto');
 const hismodel_1 = require("./../his/hismodel");
 const router = (fastify, {}, next) => {
@@ -12,14 +12,13 @@ const router = (fastify, {}, next) => {
         }
         const defaultKey = crypto.createHash('md5').update(process.env.REQUEST_KEY).digest('hex');
         if (requestKey !== defaultKey) {
-            console.log('invalid key', requestKey);
-            reply.send({
-                statusCode: HttpStatus.UNAUTHORIZED,
-                message: HttpStatus.getStatusText(HttpStatus.UNAUTHORIZED) + ' or invalid key'
+            return reply.send({
+                statusCode: http_status_codes_1.StatusCodes.UNAUTHORIZED,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.UNAUTHORIZED) + ' or invalid key'
             });
         }
-        reply.status(HttpStatus.OK).send({
-            statusCode: HttpStatus.OK,
+        reply.status(http_status_codes_1.StatusCodes.OK).send({
+            statusCode: http_status_codes_1.StatusCodes.OK,
             skey: requestKey
         });
     });
@@ -30,23 +29,23 @@ const router = (fastify, {}, next) => {
         if (searchType && searchValue) {
             try {
                 const result = await hismodel_1.default.getPerson(global.dbHIS, searchType, searchValue);
-                reply.status(HttpStatus.OK).send({
-                    statusCode: HttpStatus.OK,
+                reply.status(http_status_codes_1.StatusCodes.OK).send({
+                    statusCode: http_status_codes_1.StatusCodes.OK,
                     rows: result
                 });
             }
             catch (error) {
                 console.log('person', searchValue, error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -57,23 +56,23 @@ const router = (fastify, {}, next) => {
         if (fname + lname) {
             try {
                 const result = await hismodel_1.default.getPersonByName(global.dbHIS, fname, lname);
-                reply.status(HttpStatus.OK).send({
-                    statusCode: HttpStatus.OK,
+                reply.status(http_status_codes_1.StatusCodes.OK).send({
+                    statusCode: http_status_codes_1.StatusCodes.OK,
                     rows: result
                 });
             }
             catch (error) {
                 console.log('person', fname, lname, error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -85,30 +84,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getChronic(global.dbHIS, pid, cid);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('person-chronic', cid, pid, error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -120,30 +119,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getDrugAllergy(global.dbHIS, pid, cid);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
-                        message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                        message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
                     });
                 }
             }
             catch (error) {
                 console.log('drug-allergy', cid, pid, error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -156,20 +155,20 @@ const router = (fastify, {}, next) => {
         if (hn + cid) {
             try {
                 const rows = await hismodel_1.default.getServiceByHn(global.dbHIS, hn, cid, date, visitNo);
-                reply.status(HttpStatus.OK).send({ statusCode: HttpStatus.OK, rows });
+                reply.status(http_status_codes_1.StatusCodes.OK).send({ statusCode: http_status_codes_1.StatusCodes.OK, rows });
             }
             catch (error) {
                 console.log(error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -179,30 +178,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getDiagnosis(global.dbHIS, visitNo);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('opd-diagnosis', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -212,14 +211,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getDiagnosisByHn(global.dbHIS, hn);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -227,15 +226,15 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log('opd-diagnosis-by-hn', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -245,30 +244,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getDrug(global.dbHIS, visitNo);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('drug', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -278,14 +277,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getDrugByHn(global.dbHIS, hn);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -293,15 +292,15 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log('drug-by-hn', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -311,30 +310,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getAnc(global.dbHIS, visitNo);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('anc', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -344,14 +343,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getAncByHn(global.dbHIS, hn);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -359,15 +358,15 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log('anc-by-hn', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -377,30 +376,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getEpi(global.dbHIS, visitNo);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('epi', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -410,14 +409,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getEpiByHn(global.dbHIS, hn);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -425,15 +424,15 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log(process.env.HOSPCODE, 'epi-by-hn', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -443,30 +442,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getFp(global.dbHIS, visitNo);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('fp', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -476,14 +475,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getFpByHn(global.dbHIS, hn);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -491,15 +490,15 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log('fp-by-hn', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -509,30 +508,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getNutrition(global.dbHIS, visitNo);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('nutrition', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -542,14 +541,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getNutritionByHn(global.dbHIS, hn);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -557,15 +556,15 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log('nutrition-by-hn', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });
@@ -576,14 +575,14 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.getLabResult(global.dbHIS, searchType, searchValue);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
                     reply.send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
@@ -591,14 +590,14 @@ const router = (fastify, {}, next) => {
             catch (error) {
                 console.log('lab-result', error.message);
                 reply.send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
             reply.send({
-                statusCode: HttpStatus.BAD_REQUEST,
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                 message: 'Invalid parameter'
             });
         }
@@ -610,30 +609,30 @@ const router = (fastify, {}, next) => {
             try {
                 const result = await hismodel_1.default.libDrug(global.dbHIS, searchType, searchValue);
                 if (result) {
-                    reply.status(HttpStatus.OK).send({
-                        statusCode: HttpStatus.OK,
+                    reply.status(http_status_codes_1.StatusCodes.OK).send({
+                        statusCode: http_status_codes_1.StatusCodes.OK,
                         rows: result
                     });
                 }
                 else {
-                    reply.status(HttpStatus.BAD_REQUEST).send({
-                        statusCode: HttpStatus.BAD_REQUEST,
+                    reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                        statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
                         message: 'record not found'
                     });
                 }
             }
             catch (error) {
                 console.log('drug', error.message);
-                reply.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
-                    statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+                reply.status(http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR).send({
+                    statusCode: http_status_codes_1.StatusCodes.INTERNAL_SERVER_ERROR,
                     message: error.message
                 });
             }
         }
         else {
-            reply.status(HttpStatus.BAD_REQUEST).send({
-                statusCode: HttpStatus.BAD_REQUEST,
-                message: HttpStatus.getStatusText(HttpStatus.BAD_REQUEST)
+            reply.status(http_status_codes_1.StatusCodes.BAD_REQUEST).send({
+                statusCode: http_status_codes_1.StatusCodes.BAD_REQUEST,
+                message: (0, http_status_codes_1.getReasonPhrase)(http_status_codes_1.StatusCodes.BAD_REQUEST)
             });
         }
     });